+++ in4073/in4073.h
@@ -138,6 +138,6 @@ void ble_init(void);
 void ble_send(void);
 
 //communication check;
-void comm_check(uint16_t comm_duration, uint16_t *total_dur, bool *update_flag);
+void comm_check(uint16_t comm_duration, uint32_t *total_dur, bool *update_flag);
 
 #endif // IN4073_H__
